% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{compareWithOtherTopConditions}
\alias{compareWithOtherTopConditions}
\title{Perform analysis without reference}
\usage{
compareWithOtherTopConditions(
  mergedData,
  sampForAnalysis,
  nReads = 10,
  clones = NULL
)
}
\arguments{
\item{mergedData}{a list of data frames with read counts
for each sample.}

\item{sampForAnalysis}{a vector with sample IDs.}

\item{nReads}{a threshold for the number of reads. Default is 10.}

\item{clones}{an optional vector with clones to analyze.
Default is NULL, which means that all clones with the number of
reads more than `nReads` will be analyzed.}
}
\value{
a table with FDRs and odd ratio for comparisons with
the second and third top conditions.
}
\description{
`compareWithOtherTopConditions`
selects clones that pass the `nReads` threshold, compares
top conditions with the second and third top conditions
and returns a table with FDRs and odd ratios for those comparisons.
The output is used to select positive clones using the FDR and
odd ratio thresholds. There is no comparison to the reference
condition in this analysis.
}
