% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{createPosClonesOutput}
\alias{createPosClonesOutput}
\title{getPositiveClonesFromTopConditions}
\usage{
createPosClonesOutput(
  posClones,
  mergedData,
  refSamp = NULL,
  baselineSamp = NULL,
  addDiff = T
)
}
\description{
getPositiveClonesFromTopConditions
}
