% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{createResTable}
\alias{createResTable}
\title{createResTable}
\usage{
createResTable(
  res,
  mergedData,
  orThr = 1,
  fdrThr = 0.05,
  percentThr = 0,
  condsThr = 0,
  saveCI = T,
  significanceTable = F
)
}
\arguments{
\item{res}{a list of tables with results of Fisher's test for each pair of comparisons}

\item{mergedData}{a list of data frames with read counts for each sample}

\item{orThr}{a threshold for odds ratio}

\item{fdrThr}{a threshold for FDR}

\item{percentThr}{a threshold for the percentage of reads}

\item{condsThr}{a threshold for the percent of conditions with non-zero counts}

\item{saveCI}{a logical value indicating if confidence intervals should be saved}

\item{significanceTable}{a logical value indicating if a table with significant clones should be returned}
}
\value{
a data frame with significant clones and the corresponding
 OR, FDR, counts, and percentages for all conditions
}
\description{
Creates a table with significant clones and conditions
}
