% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{cTabPR}
\alias{cTabPR}
\title{Make matrix for one clone}
\usage{
cTabPR(clone, countData, correct = 0.5)
}
\arguments{
\item{clone}{a clone to get counts for.}

\item{countData}{a list of per clone counts for all samples
from `readMergeSave`.}

\item{correct}{a parameter to add to all counts to avoid zeros.}
}
\value{
a vector with counts for a clone across all samples.
}
\description{
Returns the count matrix for one clone from merged data.
}
