% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{getAbundances}
\alias{getAbundances}
\title{getAbundances}
\usage{
getAbundances(clones, countData)
}
\arguments{
\item{clones}{a vector of clones of interest}

\item{countData}{a list of counts for all samples}
}
\value{
a data.frame of clone abundances across samples in `countData`
}
\description{
returns the read count for clones of interest in all samples
}
