% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{getCountsPercent}
\alias{getCountsPercent}
\title{getCountsPercent}
\usage{
getCountsPercent(clones, mergedData, samp = names(mergedData))
}
\arguments{
\item{clones}{a vector with clones to analyze}

\item{mergedData}{a list of data frames with read counts for each sample}

\item{samp}{a vector with sample IDs}
}
\value{
a data frame with read counts and frequencies for selected clones and samples
}
\description{
Returns read counts and frequencies for selected clones and samples
}
