% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{getExpanded}
\alias{getExpanded}
\title{Find expanded clones}
\usage{
getExpanded(fitResults, countData, refSamp, orThr = 1, fdrThr = 0.05)
}
\arguments{
\item{fitResults}{a data frame with odd ratios, p-values, and FDRs
that outputs `fitModelSet`.}

\item{countData}{a list of counts for all samples.}

\item{refSamp}{name of reference sample.}

\item{orThr}{threshold for OR to consider a clone expanded.}

\item{fdrThr}{threshold for FDR to consider a clone expanded.}
}
\value{
a data frame with expanded clones.
}
\description{
`getExpanded` returns the expanded clones relative
to reference condition.
}
