% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{getFC}
\alias{getFC}
\title{return fold change of frequencies for selected clones and samples}
\usage{
getFC(clones, mergedData, refSamp, samp = names(mergedData), colSuf = "FC:")
}
\arguments{
\item{clones}{a vector of clones to calculate FC}

\item{mergedData}{a list of clones and
the corresponding read for all samples in the experiment}

\item{refSamp}{one sample from mergedData relative to which
the fold change should be calculated}

\item{samp}{a vector of sample names for which FC should be calculated}

\item{colSuf}{a string to be used at the beginning of columns with FC}
}
\description{
return fold change of frequencies for selected clones and samples
}
