% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{getFreq}
\alias{getFreq}
\title{getFreq}
\usage{
getFreq(
  clones,
  mergedData,
  samp = names(mergedData),
  colSuf = "percent",
  minRead = 0
)
}
\arguments{
\item{clones}{a vector with clones, for which frequencies should be calculated}

\item{mergedData}{a list of data frames with read counts for each sample}

\item{samp}{a vector with sample IDs}

\item{colSuf}{a suffix for column names}

\item{minRead}{a threshold for the number of reads}
}
\value{
a data frame with frequencies in percent for selected clones and samples
}
\description{
Returns frequencies in percent for selected clones and samples
}
