% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{getFreqThreshold}
\alias{getFreqThreshold}
\title{getFreqThreshold}
\usage{
getFreqThreshold(n, p)
}
\arguments{
\item{n}{number of cells per well}

\item{p}{selected probability}
}
\value{
frequency threshold
}
\description{
Calculates frequency threshold using the number of cells and probability
(1-((1-p)^(1/n)))
where n=number of cells per well and p=selected probability
}
