% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{getPositiveClonesFromTopConditions}
\alias{getPositiveClonesFromTopConditions}
\title{getPositiveClonesFromTopConditions}
\usage{
getPositiveClonesFromTopConditions(fisherResTable, orThr = 1, fdrThr = 0.05)
}
\description{
getPositiveClonesFromTopConditions
}
