% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{getPositiveClonesReplicates}
\alias{getPositiveClonesReplicates}
\title{getPositiveClonesReplicates
returns positive clones for version with replicates}
\usage{
getPositiveClonesReplicates(
  analysisRes,
  mergedData,
  control,
  samp = names(mergedData),
  excludeCond = NA,
  orThr = 1,
  fdrThr = 0.05,
  percentThr = 0
)
}
\arguments{
\item{analysisRes}{a data frame with results of analysis}

\item{mergedData}{a list of data frames with read counts for each sample}

\item{control}{a name of control condition to compare to}

\item{samp}{a vector of sample names to include in the analysis}

\item{excludeCond}{a vector of conditions to exclude from the analysis}

\item{orThr}{a threshold for odds ratio}

\item{fdrThr}{a threshold for FDR}

\item{percentThr}{a threshold for percentage of reads in a sample to consider a clone expanded}
}
\value{
a data frame with positive clones, significant condition,
and significant sample
}
\description{
getPositiveClonesReplicates
returns positive clones for version with replicates
}
