% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{getXR}
\alias{getXR}
\title{Returns the cross-reactive clones}
\usage{
getXR(res, conditions, xrCond)
}
\arguments{
\item{res}{a data frame with odd ratios, p-values, and FDRs
that outputs `createResTableReplicates`.}

\item{conditions}{a vector of all conditions that were used in the analysis}

\item{xrCond}{a vector of conditions that could be cross-reactive}

\item{excludeCond}{a vector of conditions that were excluded from the analysis}
}
\value{
a data frame with cross-reactive clones.
}
\description{
`getXR` returns output with cross-reactive clones in specified
conditions, if any.
}
