% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{makeHeatmaps}
\alias{makeHeatmaps}
\title{Plot heatmap of frequencies or fold changes}
\usage{
makeHeatmaps(
  clones,
  mergedData,
  refSamp = NULL,
  fileName = "heatmap.pdf",
  size = 7
)
}
\arguments{
\item{clones}{a vector of clones to plot.}

\item{mergedData}{a list of data frames with read counts
for each sample.}

\item{refSamp}{a reference sample. If NULL,
then a heatmap of frequencies is plotted.}

\item{fileName}{a name of the output PDF file.}

\item{size}{a height and width of heatmap in the output PDF file.}
}
\description{
Makes heatmap of frequencies or fold changes
(if `refSamp` is specified) for the specified clones and samples.
}
