% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions-EB-ONC-60W6DZ3.r,
%   R/repManFunctions.r
\name{readMergeSave}
\alias{readMergeSave}
\title{readMergeSave}
\usage{
readMergeSave(files, filenames = NULL)

readMergeSave(files, filenames = NULL)
}
\arguments{
\item{files}{a list of filenames with full paths}

\item{filenames}{a vector of filenames}
}
\value{
a list with mergedData, ntData, productiveReadCounts

a list of AA level counts, nucleotide level counts, and total number of reads
}
\description{
reads files, removes non-productive sequencies, extracts counts,
creates all necessary objects, and saves them
}
\examples{
readMergeSave(files = "path/to/files")

readMergeSave(files = c("data/0.csv","data/1.csv","data/2.csv"))

}
