% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicateFest-package.R
\docType{package}
\name{replicateFest-package}
\alias{replicateFest}
\alias{replicateFest-package}
\title{replicateFest: Analyze TCR-seq Data From FEST Assay}
\description{
The package contains functions to analyze TCR-sequencing data from FEST (Functional Expansion of Specific T-cells) assay. TCR sequencing data is generated from short-term, peptide-stimulated cultures to identify antigen-specific clonotypic amplifications.
}
\author{
\strong{Maintainer}: Ludmila Danilova \email{ldanilo1@jhmi.edu}

Authors:
\itemize{
  \item Leslie Cope \email{lcope1@jhmi.edu}
  \item Alexander Favorov \email{avf@jhmi.edu}
}

}
\keyword{internal}
