% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{runExperiment}
\alias{runExperiment}
\title{runExperiment}
\usage{
runExperiment(
  files,
  peptides,
  ctThresh = 50,
  control,
  ORthr = 1,
  FDRthr = 0.05,
  excludeCond = NA,
  xrCond = NA,
  percentThr = 0,
  outputFile = "output.xlsx",
  saveToFile = T,
  permute = FALSE
)
}
\arguments{
\item{files}{a list of filenames with full paths}

\item{peptides}{a vector of peptides corresponding to columns in merged data}

\item{ctThresh}{minimal number of reads required to consider a clone}

\item{control}{name of control/reference condition}

\item{ORthr}{threshold for OR to consider a clone expanded}

\item{FDRthr}{threshold for FDR to consider a clone expanded}

\item{excludeCond}{a vector of conditions to exclude from the analysis}

\item{xrCond}{a vector of cross-reactive conditions}

\item{percentThr}{a threshold for percentage of reads in a sample to consider a clone expanded}

\item{outputFile}{name of the output file}

\item{saveToFile}{logical, if TRUE save results to a file}

\item{permute}{logical, if TRUE permute sample labels to run a permutation test}
}
\value{
a list of all expanded clones, uniquely expanded clones
and parameters of the run
}
\description{
Reads in files with TCR repertoires from
a FEST experiment with replicate samples per condition (stimulating peptide).
It fits negative binomial model to find expanded clones
comparing to a reference samples.
It also compares top conditions to find unique expansions.
The results are return and saved in an Excel file.
}
