% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{saveResults}
\alias{saveResults}
\title{Save results to an Excel file}
\usage{
saveResults(results, outputFile = "output.xlsx")
}
\arguments{
\item{results}{a list of data frames with results.}

\item{outputFile}{name of the output file.}
}
\description{
`saveResults` saves results to an excel file.
}
\details{
saveResults
}
