% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{splitFileName}
\alias{splitFileName}
\title{function that extracts condition and replicate information
from the file names. The condition and replicate should
be separated by "_" and be the last two elements.}
\usage{
splitFileName(filenames)
}
\arguments{
\item{filenames}{a vector of file names}
}
\value{
a data frame with condition and replicate information
}
\description{
function that extracts condition and replicate information
from the file names. The condition and replicate should
be separated by "_" and be the last two elements.
}
