% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{compareWithOtherTopConditions}
\alias{compareWithOtherTopConditions}
\title{compareWithOtherTopConditions}
\usage{
compareWithOtherTopConditions(
  mergedData,
  sampForAnalysis,
  nReads = 10,
  clones = NULL
)
}
\description{
compareWithOtherTopConditions
}
