% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{createPosClonesOutput}
\alias{createPosClonesOutput}
\title{createPosClonesOutput}
\usage{
createPosClonesOutput(
  posClones,
  mergedData,
  refSamp = NULL,
  replicates = FALSE
)
}
\arguments{
\item{posClones}{a data frame with positive clones in the first column
and their conditions to be summarized in the second.
In the replication version, these should be samples}

\item{mergedData}{a list of data frames with read counts for each sample}

\item{refSamp}{a reference sample ID. If NULL, then FC columns will not be added}

\item{replicates}{a logical value indicating if there are replicates}
}
\value{
a list with two data frames:
1) condition_summary - a table with the number of positive clones per sample
and the sum of their frequencies
2) positive_clones_summary - a table with per clone information
including frequencies and fold changes relative to the reference sample
}
\description{
Creates output tables to be saved in Excel
}
