% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{cTabPR}
\alias{cTabPR}
\title{cTabPR
function to make the count matrix for 1 clone, from merged data}
\usage{
cTabPR(clone, countData, correct = 1)
}
\arguments{
\item{clone}{a clone to get counts for}

\item{countData}{a list of per clone counts for all samples from readMergeSave}

\item{correct}{a parameter to add to all counts to avoid 0s}
}
\value{
a vector with counts for a clone across all samples
}
\description{
cTabPR
function to make the count matrix for 1 clone, from merged data
}
