% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{fitModelSet}
\alias{fitModelSet}
\title{fit model for a set of clones}
\usage{
fitModelSet(clones, countData, peptides, excludeCond = NA, ...)
}
\arguments{
\item{clones}{a list of clones to fit the model}

\item{countData}{a list of counts for all samples}

\item{peptides}{a vector of peptides corresponding to columns in merged data}

\item{excludeCond}{a vector of conditions to exclude from the analysis}

\item{...}{additional parameters to pass to fitModel}
}
\value{
a matrix with all ORs, p-values and FDRs
}
\description{
fit model for a set of clones
}
