% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{getClonesToTest}
\alias{getClonesToTest}
\title{getClonesToTest}
\usage{
getClonesToTest(countDat, nReads = 50)
}
\arguments{
\item{countDat}{a list of merged data}

\item{nReads}{a minimal number of reads required to consider a clone}
}
\value{
a vector of clones of interest
}
\description{
function that returns clones of interest in all samples
}
