% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{getPositiveClones}
\alias{getPositiveClones}
\title{getPositiveClones}
\usage{
getPositiveClones(
  analysisRes,
  mergedData,
  samp = names(mergedData),
  orThr = 1,
  fdrThr = 0.05,
  nReads = 10
)
}
\description{
getPositiveClones
}
