% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manafest_shiny_functions.r
\name{makeHeatmaps}
\alias{makeHeatmaps}
\title{makeHeatmaps}
\usage{
makeHeatmaps(
  listOfclones,
  mergedData,
  samp = names(mergedData),
  refSamp = NULL,
  fileName = "heatmap.pdf",
  size = 7
)
}
\arguments{
\item{listOfclones}{a list of clones to plot}

\item{mergedData}{a list of data frames with read counts for each sample}

\item{samp}{a vector with sample IDs}

\item{refSamp}{a reference sample ID}

\item{fileName}{a name of the output file}

\item{size}{a height and width of heatmap in the output PDF file}
}
\description{
Makes heatmaps of frequencies of each element
of input list of clones and samples
and plots fold change if refSamp is specified
}
