% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{readMergeSave}
\alias{readMergeSave}
\title{readMergeSave
Reads files, removes non-productive sequences, extracts counts,
creates all necessary objects for further analysis}
\usage{
readMergeSave(files, filenames = NULL)
}
\arguments{
\item{files}{a list of filenames with full paths}

\item{filenames}{a vector of filenames}
}
\value{
a list of AA level counts, nucleotide level counts, and total number of reads
}
\description{
readMergeSave
Reads files, removes non-productive sequences, extracts counts,
creates all necessary objects for further analysis
}
\examples{
readMergeSave(files = c("data/0.csv","data/1.csv","data/2.csv"))

}
