% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repManFunctions.r
\name{saveResults}
\alias{saveResults}
\title{saveResults}
\usage{
saveResults(results, outputFile = "output.xlsx")
}
\arguments{
\item{results}{a list of data frames with results}

\item{outputFile}{name of the output file}
}
\description{
Saves results to an excel file
}
\details{
saveResults
}
